+++ CHANGELOG.md
@@ -42,6 +42,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-1118](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1118)
 - Move binning by season creation to a separate endpoint 
   [GEOD-1153](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1153)
+- Changed Binning for stacked bar data, implemented only on season bins
+  [GEOD-1155](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1155)
  
 ### Fixed 
 - Out of memory error when user requests large downloads